Refactor Xml to extract transformer creation and Fix broken test
